tests(configHelper): Convert to Jasmine Tests
